requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
convertdate==2.4.0
Django==4.2.15
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dotmap==1.3.30
exceptiongroup==1.2.0
frozenlist==1.4.1
func-timeout==4.3.5
gunicorn==23.0.0
h11==0.14.0
idna==3.6
lunardate==0.2.2
multidict==6.0.5
mysqlclient==2.0.3
outcome==1.3.0.post0
packaging==24.0
pillow==10.4.0
py-cord==2.5.0
pycparser==2.22
pyluach==2.2.0
python-decouple==3.8
PyMeeus==0.5.12
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.10.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
webdriver-manager==4.0.1
websocket-client==1.8.0
workalendar==17.0.0
wsproto==1.2.0
yarl==1.9.4
